Add RecommendedActivities component
